app/requirements.txt
fastapi>=0.95.0
python-multipart>=0.0.6
ffmpeg-python>=0.2.0
uvicorn>=0.22.0
